requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.9
decorator==5.1.1
duckdb==1.1.3
et_xmlfile==2.0.0
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msoffcrypto-tool==5.4.2
narwhals==1.14.3
nest-asyncio==1.6.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sql_metadata==2.15.0
sqlparse==0.5.2
st-gsheets-connection==0.1.0
stack-data==0.6.3
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
wcwidth==0.2.13
xlrd==2.0.1
